8293230: x86_64: Move AES and GHASH stub definitions into separate source files
