+++ src/hotspot/cpu/x86/stubGenerator_x86_64.cpp
@@ -1259,683 +1259,6 @@ address StubGenerator::generate_data_cache_writeback_sync() {
   return start;
 }
 
-// AES intrinsic stubs
-
-address StubGenerator::generate_key_shuffle_mask() {
-  __ align(16);
-  StubCodeMark mark(this, "StubRoutines", "key_shuffle_mask");
-  address start = __ pc();
-
-  __ emit_data64( 0x0405060700010203, relocInfo::none );
-  __ emit_data64( 0x0c0d0e0f08090a0b, relocInfo::none );
-
-  return start;
-}
-
-address StubGenerator::generate_counter_shuffle_mask() {
-  __ align(16);
-  StubCodeMark mark(this, "StubRoutines", "counter_shuffle_mask");
-  address start = __ pc();
-
-  __ emit_data64(0x08090a0b0c0d0e0f, relocInfo::none);
-  __ emit_data64(0x0001020304050607, relocInfo::none);
-
-  return start;
-}
-
-// Utility routine for loading a 128-bit key word in little endian format
-// can optionally specify that the shuffle mask is already in an xmmregister
-void StubGenerator::load_key(XMMRegister xmmdst, Register key, int offset, XMMRegister xmm_shuf_mask) {
-  __ movdqu(xmmdst, Address(key, offset));
-  if (xmm_shuf_mask != xnoreg) {
-    __ pshufb(xmmdst, xmm_shuf_mask);
-  } else {
-    __ pshufb(xmmdst, ExternalAddress(StubRoutines::x86::key_shuffle_mask_addr()));
-  }
-}
-
-// Utility routine for increase 128bit counter (iv in CTR mode)
-void StubGenerator::inc_counter(Register reg, XMMRegister xmmdst, int inc_delta, Label& next_block) {
-  __ pextrq(reg, xmmdst, 0x0);
-  __ addq(reg, inc_delta);
-  __ pinsrq(xmmdst, reg, 0x0);
-  __ jcc(Assembler::carryClear, next_block); // jump if no carry
-  __ pextrq(reg, xmmdst, 0x01); // Carry
-  __ addq(reg, 0x01);
-  __ pinsrq(xmmdst, reg, 0x01); //Carry end
-  __ BIND(next_block);          // next instruction
-}
-
-// Arguments:
-//
-// Inputs:
-//   c_rarg0   - source byte array address
-//   c_rarg1   - destination byte array address
-//   c_rarg2   - K (key) in little endian int array
-//
-address StubGenerator::generate_aescrypt_encryptBlock() {
-  assert(UseAES, "need AES instructions and misaligned SSE support");
-  __ align(CodeEntryAlignment);
-  StubCodeMark mark(this, "StubRoutines", "aescrypt_encryptBlock");
-  Label L_doLast;
-  address start = __ pc();
-
-  const Register from        = c_rarg0;  // source array address
-  const Register to          = c_rarg1;  // destination array address
-  const Register key         = c_rarg2;  // key array address
-  const Register keylen      = rax;
-
-  const XMMRegister xmm_result = xmm0;
-  const XMMRegister xmm_key_shuf_mask = xmm1;
-  // On win64 xmm6-xmm15 must be preserved so don't use them.
-  const XMMRegister xmm_temp1  = xmm2;
-  const XMMRegister xmm_temp2  = xmm3;
-  const XMMRegister xmm_temp3  = xmm4;
-  const XMMRegister xmm_temp4  = xmm5;
-
-  __ enter(); // required for proper stackwalking of RuntimeStub frame
-
-  // keylen could be only {11, 13, 15} * 4 = {44, 52, 60}
-  __ movl(keylen, Address(key, arrayOopDesc::length_offset_in_bytes() - arrayOopDesc::base_offset_in_bytes(T_INT)));
-
-  __ movdqu(xmm_key_shuf_mask, ExternalAddress(StubRoutines::x86::key_shuffle_mask_addr()));
-  __ movdqu(xmm_result, Address(from, 0));  // get 16 bytes of input
-
-  // For encryption, the java expanded key ordering is just what we need
-  // we don't know if the key is aligned, hence not using load-execute form
-
-  load_key(xmm_temp1, key, 0x00, xmm_key_shuf_mask);
-  __ pxor(xmm_result, xmm_temp1);
-
-  load_key(xmm_temp1, key, 0x10, xmm_key_shuf_mask);
-  load_key(xmm_temp2, key, 0x20, xmm_key_shuf_mask);
-  load_key(xmm_temp3, key, 0x30, xmm_key_shuf_mask);
-  load_key(xmm_temp4, key, 0x40, xmm_key_shuf_mask);
-
-  __ aesenc(xmm_result, xmm_temp1);
-  __ aesenc(xmm_result, xmm_temp2);
-  __ aesenc(xmm_result, xmm_temp3);
-  __ aesenc(xmm_result, xmm_temp4);
-
-  load_key(xmm_temp1, key, 0x50, xmm_key_shuf_mask);
-  load_key(xmm_temp2, key, 0x60, xmm_key_shuf_mask);
-  load_key(xmm_temp3, key, 0x70, xmm_key_shuf_mask);
-  load_key(xmm_temp4, key, 0x80, xmm_key_shuf_mask);
-
-  __ aesenc(xmm_result, xmm_temp1);
-  __ aesenc(xmm_result, xmm_temp2);
-  __ aesenc(xmm_result, xmm_temp3);
-  __ aesenc(xmm_result, xmm_temp4);
-
-  load_key(xmm_temp1, key, 0x90, xmm_key_shuf_mask);
-  load_key(xmm_temp2, key, 0xa0, xmm_key_shuf_mask);
-
-  __ cmpl(keylen, 44);
-  __ jccb(Assembler::equal, L_doLast);
-
-  __ aesenc(xmm_result, xmm_temp1);
-  __ aesenc(xmm_result, xmm_temp2);
-
-  load_key(xmm_temp1, key, 0xb0, xmm_key_shuf_mask);
-  load_key(xmm_temp2, key, 0xc0, xmm_key_shuf_mask);
-
-  __ cmpl(keylen, 52);
-  __ jccb(Assembler::equal, L_doLast);
-
-  __ aesenc(xmm_result, xmm_temp1);
-  __ aesenc(xmm_result, xmm_temp2);
-
-  load_key(xmm_temp1, key, 0xd0, xmm_key_shuf_mask);
-  load_key(xmm_temp2, key, 0xe0, xmm_key_shuf_mask);
-
-  __ BIND(L_doLast);
-  __ aesenc(xmm_result, xmm_temp1);
-  __ aesenclast(xmm_result, xmm_temp2);
-  __ movdqu(Address(to, 0), xmm_result);        // store the result
-  __ xorptr(rax, rax); // return 0
-  __ leave(); // required for proper stackwalking of RuntimeStub frame
-  __ ret(0);
-
-  return start;
-}
-
-
-// Arguments:
-//
-// Inputs:
-//   c_rarg0   - source byte array address
-//   c_rarg1   - destination byte array address
-//   c_rarg2   - K (key) in little endian int array
-//
-address StubGenerator::generate_aescrypt_decryptBlock() {
-  assert(UseAES, "need AES instructions and misaligned SSE support");
-  __ align(CodeEntryAlignment);
-  StubCodeMark mark(this, "StubRoutines", "aescrypt_decryptBlock");
-  Label L_doLast;
-  address start = __ pc();
-
-  const Register from        = c_rarg0;  // source array address
-  const Register to          = c_rarg1;  // destination array address
-  const Register key         = c_rarg2;  // key array address
-  const Register keylen      = rax;
-
-  const XMMRegister xmm_result = xmm0;
-  const XMMRegister xmm_key_shuf_mask = xmm1;
-  // On win64 xmm6-xmm15 must be preserved so don't use them.
-  const XMMRegister xmm_temp1  = xmm2;
-  const XMMRegister xmm_temp2  = xmm3;
-  const XMMRegister xmm_temp3  = xmm4;
-  const XMMRegister xmm_temp4  = xmm5;
-
-  __ enter(); // required for proper stackwalking of RuntimeStub frame
-
-  // keylen could be only {11, 13, 15} * 4 = {44, 52, 60}
-  __ movl(keylen, Address(key, arrayOopDesc::length_offset_in_bytes() - arrayOopDesc::base_offset_in_bytes(T_INT)));
-
-  __ movdqu(xmm_key_shuf_mask, ExternalAddress(StubRoutines::x86::key_shuffle_mask_addr()));
-  __ movdqu(xmm_result, Address(from, 0));
-
-  // for decryption java expanded key ordering is rotated one position from what we want
-  // so we start from 0x10 here and hit 0x00 last
-  // we don't know if the key is aligned, hence not using load-execute form
-  load_key(xmm_temp1, key, 0x10, xmm_key_shuf_mask);
-  load_key(xmm_temp2, key, 0x20, xmm_key_shuf_mask);
-  load_key(xmm_temp3, key, 0x30, xmm_key_shuf_mask);
-  load_key(xmm_temp4, key, 0x40, xmm_key_shuf_mask);
-
-  __ pxor  (xmm_result, xmm_temp1);
-  __ aesdec(xmm_result, xmm_temp2);
-  __ aesdec(xmm_result, xmm_temp3);
-  __ aesdec(xmm_result, xmm_temp4);
-
-  load_key(xmm_temp1, key, 0x50, xmm_key_shuf_mask);
-  load_key(xmm_temp2, key, 0x60, xmm_key_shuf_mask);
-  load_key(xmm_temp3, key, 0x70, xmm_key_shuf_mask);
-  load_key(xmm_temp4, key, 0x80, xmm_key_shuf_mask);
-
-  __ aesdec(xmm_result, xmm_temp1);
-  __ aesdec(xmm_result, xmm_temp2);
-  __ aesdec(xmm_result, xmm_temp3);
-  __ aesdec(xmm_result, xmm_temp4);
-
-  load_key(xmm_temp1, key, 0x90, xmm_key_shuf_mask);
-  load_key(xmm_temp2, key, 0xa0, xmm_key_shuf_mask);
-  load_key(xmm_temp3, key, 0x00, xmm_key_shuf_mask);
-
-  __ cmpl(keylen, 44);
-  __ jccb(Assembler::equal, L_doLast);
-
-  __ aesdec(xmm_result, xmm_temp1);
-  __ aesdec(xmm_result, xmm_temp2);
-
-  load_key(xmm_temp1, key, 0xb0, xmm_key_shuf_mask);
-  load_key(xmm_temp2, key, 0xc0, xmm_key_shuf_mask);
-
-  __ cmpl(keylen, 52);
-  __ jccb(Assembler::equal, L_doLast);
-
-  __ aesdec(xmm_result, xmm_temp1);
-  __ aesdec(xmm_result, xmm_temp2);
-
-  load_key(xmm_temp1, key, 0xd0, xmm_key_shuf_mask);
-  load_key(xmm_temp2, key, 0xe0, xmm_key_shuf_mask);
-
-  __ BIND(L_doLast);
-  __ aesdec(xmm_result, xmm_temp1);
-  __ aesdec(xmm_result, xmm_temp2);
-
-  // for decryption the aesdeclast operation is always on key+0x00
-  __ aesdeclast(xmm_result, xmm_temp3);
-  __ movdqu(Address(to, 0), xmm_result);  // store the result
-  __ xorptr(rax, rax); // return 0
-  __ leave(); // required for proper stackwalking of RuntimeStub frame
-  __ ret(0);
-
-  return start;
-}
-
-
-// Arguments:
-//
-// Inputs:
-//   c_rarg0   - source byte array address
-//   c_rarg1   - destination byte array address
-//   c_rarg2   - K (key) in little endian int array
-//   c_rarg3   - r vector byte array address
-//   c_rarg4   - input length
-//
-// Output:
-//   rax       - input length
-//
-address StubGenerator::generate_cipherBlockChaining_encryptAESCrypt() {
-  assert(UseAES, "need AES instructions and misaligned SSE support");
-  __ align(CodeEntryAlignment);
-  StubCodeMark mark(this, "StubRoutines", "cipherBlockChaining_encryptAESCrypt");
-  address start = __ pc();
-
-  Label L_exit, L_key_192_256, L_key_256, L_loopTop_128, L_loopTop_192, L_loopTop_256;
-  const Register from        = c_rarg0;  // source array address
-  const Register to          = c_rarg1;  // destination array address
-  const Register key         = c_rarg2;  // key array address
-  const Register rvec        = c_rarg3;  // r byte array initialized from initvector array address
-                                         // and left with the results of the last encryption block
-#ifndef _WIN64
-  const Register len_reg     = c_rarg4;  // src len (must be multiple of blocksize 16)
-#else
-  const Address  len_mem(rbp, 6 * wordSize);  // length is on stack on Win64
-  const Register len_reg     = r11;      // pick the volatile windows register
-#endif
-  const Register pos         = rax;
-
-  // xmm register assignments for the loops below
-  const XMMRegister xmm_result = xmm0;
-  const XMMRegister xmm_temp   = xmm1;
-  // keys 0-10 preloaded into xmm2-xmm12
-  const int XMM_REG_NUM_KEY_FIRST = 2;
-  const int XMM_REG_NUM_KEY_LAST  = 15;
-  const XMMRegister xmm_key0   = as_XMMRegister(XMM_REG_NUM_KEY_FIRST);
-  const XMMRegister xmm_key10  = as_XMMRegister(XMM_REG_NUM_KEY_FIRST+10);
-  const XMMRegister xmm_key11  = as_XMMRegister(XMM_REG_NUM_KEY_FIRST+11);
-  const XMMRegister xmm_key12  = as_XMMRegister(XMM_REG_NUM_KEY_FIRST+12);
-  const XMMRegister xmm_key13  = as_XMMRegister(XMM_REG_NUM_KEY_FIRST+13);
-
-  __ enter(); // required for proper stackwalking of RuntimeStub frame
-
-#ifdef _WIN64
-  // on win64, fill len_reg from stack position
-  __ movl(len_reg, len_mem);
-#else
-  __ push(len_reg); // Save
-#endif
-
-  const XMMRegister xmm_key_shuf_mask = xmm_temp;  // used temporarily to swap key bytes up front
-  __ movdqu(xmm_key_shuf_mask, ExternalAddress(StubRoutines::x86::key_shuffle_mask_addr()));
-  // load up xmm regs xmm2 thru xmm12 with key 0x00 - 0xa0
-  for (int rnum = XMM_REG_NUM_KEY_FIRST, offset = 0x00; rnum <= XMM_REG_NUM_KEY_FIRST+10; rnum++) {
-    load_key(as_XMMRegister(rnum), key, offset, xmm_key_shuf_mask);
-    offset += 0x10;
-  }
-  __ movdqu(xmm_result, Address(rvec, 0x00));   // initialize xmm_result with r vec
-
-  // now split to different paths depending on the keylen (len in ints of AESCrypt.KLE array (52=192, or 60=256))
-  __ movl(rax, Address(key, arrayOopDesc::length_offset_in_bytes() - arrayOopDesc::base_offset_in_bytes(T_INT)));
-  __ cmpl(rax, 44);
-  __ jcc(Assembler::notEqual, L_key_192_256);
-
-  // 128 bit code follows here
-  __ movptr(pos, 0);
-  __ align(OptoLoopAlignment);
-
-  __ BIND(L_loopTop_128);
-  __ movdqu(xmm_temp, Address(from, pos, Address::times_1, 0));   // get next 16 bytes of input
-  __ pxor  (xmm_result, xmm_temp);               // xor with the current r vector
-  __ pxor  (xmm_result, xmm_key0);               // do the aes rounds
-  for (int rnum = XMM_REG_NUM_KEY_FIRST + 1; rnum <= XMM_REG_NUM_KEY_FIRST + 9; rnum++) {
-    __ aesenc(xmm_result, as_XMMRegister(rnum));
-  }
-  __ aesenclast(xmm_result, xmm_key10);
-  __ movdqu(Address(to, pos, Address::times_1, 0), xmm_result);     // store into the next 16 bytes of output
-  // no need to store r to memory until we exit
-  __ addptr(pos, AESBlockSize);
-  __ subptr(len_reg, AESBlockSize);
-  __ jcc(Assembler::notEqual, L_loopTop_128);
-
-  __ BIND(L_exit);
-  __ movdqu(Address(rvec, 0), xmm_result);     // final value of r stored in rvec of CipherBlockChaining object
-
-#ifdef _WIN64
-  __ movl(rax, len_mem);
-#else
-  __ pop(rax); // return length
-#endif
-  __ leave(); // required for proper stackwalking of RuntimeStub frame
-  __ ret(0);
-
-  __ BIND(L_key_192_256);
-  // here rax = len in ints of AESCrypt.KLE array (52=192, or 60=256)
-  load_key(xmm_key11, key, 0xb0, xmm_key_shuf_mask);
-  load_key(xmm_key12, key, 0xc0, xmm_key_shuf_mask);
-  __ cmpl(rax, 52);
-  __ jcc(Assembler::notEqual, L_key_256);
-
-  // 192-bit code follows here (could be changed to use more xmm registers)
-  __ movptr(pos, 0);
-  __ align(OptoLoopAlignment);
-
-  __ BIND(L_loopTop_192);
-  __ movdqu(xmm_temp, Address(from, pos, Address::times_1, 0));   // get next 16 bytes of input
-  __ pxor  (xmm_result, xmm_temp);               // xor with the current r vector
-  __ pxor  (xmm_result, xmm_key0);               // do the aes rounds
-  for (int rnum = XMM_REG_NUM_KEY_FIRST + 1; rnum  <= XMM_REG_NUM_KEY_FIRST + 11; rnum++) {
-    __ aesenc(xmm_result, as_XMMRegister(rnum));
-  }
-  __ aesenclast(xmm_result, xmm_key12);
-  __ movdqu(Address(to, pos, Address::times_1, 0), xmm_result);     // store into the next 16 bytes of output
-  // no need to store r to memory until we exit
-  __ addptr(pos, AESBlockSize);
-  __ subptr(len_reg, AESBlockSize);
-  __ jcc(Assembler::notEqual, L_loopTop_192);
-  __ jmp(L_exit);
-
-  __ BIND(L_key_256);
-  // 256-bit code follows here (could be changed to use more xmm registers)
-  load_key(xmm_key13, key, 0xd0, xmm_key_shuf_mask);
-  __ movptr(pos, 0);
-  __ align(OptoLoopAlignment);
-
-  __ BIND(L_loopTop_256);
-  __ movdqu(xmm_temp, Address(from, pos, Address::times_1, 0));   // get next 16 bytes of input
-  __ pxor  (xmm_result, xmm_temp);               // xor with the current r vector
-  __ pxor  (xmm_result, xmm_key0);               // do the aes rounds
-  for (int rnum = XMM_REG_NUM_KEY_FIRST + 1; rnum  <= XMM_REG_NUM_KEY_FIRST + 13; rnum++) {
-    __ aesenc(xmm_result, as_XMMRegister(rnum));
-  }
-  load_key(xmm_temp, key, 0xe0);
-  __ aesenclast(xmm_result, xmm_temp);
-  __ movdqu(Address(to, pos, Address::times_1, 0), xmm_result);     // store into the next 16 bytes of output
-  // no need to store r to memory until we exit
-  __ addptr(pos, AESBlockSize);
-  __ subptr(len_reg, AESBlockSize);
-  __ jcc(Assembler::notEqual, L_loopTop_256);
-  __ jmp(L_exit);
-
-  return start;
-}
-
-// This is a version of CBC/AES Decrypt which does 4 blocks in a loop at a time
-// to hide instruction latency
-//
-// Arguments:
-//
-// Inputs:
-//   c_rarg0   - source byte array address
-//   c_rarg1   - destination byte array address
-//   c_rarg2   - K (key) in little endian int array
-//   c_rarg3   - r vector byte array address
-//   c_rarg4   - input length
-//
-// Output:
-//   rax       - input length
-//
-address StubGenerator::generate_cipherBlockChaining_decryptAESCrypt_Parallel() {
-  assert(UseAES, "need AES instructions and misaligned SSE support");
-  __ align(CodeEntryAlignment);
-  StubCodeMark mark(this, "StubRoutines", "cipherBlockChaining_decryptAESCrypt");
-  address start = __ pc();
-
-  const Register from        = c_rarg0;  // source array address
-  const Register to          = c_rarg1;  // destination array address
-  const Register key         = c_rarg2;  // key array address
-  const Register rvec        = c_rarg3;  // r byte array initialized from initvector array address
-                                         // and left with the results of the last encryption block
-#ifndef _WIN64
-  const Register len_reg     = c_rarg4;  // src len (must be multiple of blocksize 16)
-#else
-  const Address  len_mem(rbp, 6 * wordSize);  // length is on stack on Win64
-  const Register len_reg     = r11;      // pick the volatile windows register
-#endif
-  const Register pos         = rax;
-
-  const int PARALLEL_FACTOR = 4;
-  const int ROUNDS[3] = { 10, 12, 14 }; // aes rounds for key128, key192, key256
-
-  Label L_exit;
-  Label L_singleBlock_loopTopHead[3]; // 128, 192, 256
-  Label L_singleBlock_loopTopHead2[3]; // 128, 192, 256
-  Label L_singleBlock_loopTop[3]; // 128, 192, 256
-  Label L_multiBlock_loopTopHead[3]; // 128, 192, 256
-  Label L_multiBlock_loopTop[3]; // 128, 192, 256
-
-  // keys 0-10 preloaded into xmm5-xmm15
-  const int XMM_REG_NUM_KEY_FIRST = 5;
-  const int XMM_REG_NUM_KEY_LAST  = 15;
-  const XMMRegister xmm_key_first = as_XMMRegister(XMM_REG_NUM_KEY_FIRST);
-  const XMMRegister xmm_key_last  = as_XMMRegister(XMM_REG_NUM_KEY_LAST);
-
-  __ enter(); // required for proper stackwalking of RuntimeStub frame
-
-#ifdef _WIN64
-  // on win64, fill len_reg from stack position
-  __ movl(len_reg, len_mem);
-#else
-  __ push(len_reg); // Save
-#endif
-  __ push(rbx);
-  // the java expanded key ordering is rotated one position from what we want
-  // so we start from 0x10 here and hit 0x00 last
-  const XMMRegister xmm_key_shuf_mask = xmm1;  // used temporarily to swap key bytes up front
-  __ movdqu(xmm_key_shuf_mask, ExternalAddress(StubRoutines::x86::key_shuffle_mask_addr()));
-  // load up xmm regs 5 thru 15 with key 0x10 - 0xa0 - 0x00
-  for (int rnum = XMM_REG_NUM_KEY_FIRST, offset = 0x10; rnum < XMM_REG_NUM_KEY_LAST; rnum++) {
-    load_key(as_XMMRegister(rnum), key, offset, xmm_key_shuf_mask);
-    offset += 0x10;
-  }
-  load_key(xmm_key_last, key, 0x00, xmm_key_shuf_mask);
-
-  const XMMRegister xmm_prev_block_cipher = xmm1;  // holds cipher of previous block
-
-  // registers holding the four results in the parallelized loop
-  const XMMRegister xmm_result0 = xmm0;
-  const XMMRegister xmm_result1 = xmm2;
-  const XMMRegister xmm_result2 = xmm3;
-  const XMMRegister xmm_result3 = xmm4;
-
-  __ movdqu(xmm_prev_block_cipher, Address(rvec, 0x00));   // initialize with initial rvec
-
-  __ xorptr(pos, pos);
-
-  // now split to different paths depending on the keylen (len in ints of AESCrypt.KLE array (52=192, or 60=256))
-  __ movl(rbx, Address(key, arrayOopDesc::length_offset_in_bytes() - arrayOopDesc::base_offset_in_bytes(T_INT)));
-  __ cmpl(rbx, 52);
-  __ jcc(Assembler::equal, L_multiBlock_loopTopHead[1]);
-  __ cmpl(rbx, 60);
-  __ jcc(Assembler::equal, L_multiBlock_loopTopHead[2]);
-
-#define DoFour(opc, src_reg)           \
-__ opc(xmm_result0, src_reg);         \
-__ opc(xmm_result1, src_reg);         \
-__ opc(xmm_result2, src_reg);         \
-__ opc(xmm_result3, src_reg);         \
-
-  for (int k = 0; k < 3; ++k) {
-    __ BIND(L_multiBlock_loopTopHead[k]);
-    if (k != 0) {
-      __ cmpptr(len_reg, PARALLEL_FACTOR * AESBlockSize); // see if at least 4 blocks left
-      __ jcc(Assembler::less, L_singleBlock_loopTopHead2[k]);
-    }
-    if (k == 1) {
-      __ subptr(rsp, 6 * wordSize);
-      __ movdqu(Address(rsp, 0), xmm15); //save last_key from xmm15
-      load_key(xmm15, key, 0xb0); // 0xb0; 192-bit key goes up to 0xc0
-      __ movdqu(Address(rsp, 2 * wordSize), xmm15);
-      load_key(xmm1, key, 0xc0);  // 0xc0;
-      __ movdqu(Address(rsp, 4 * wordSize), xmm1);
-    } else if (k == 2) {
-      __ subptr(rsp, 10 * wordSize);
-      __ movdqu(Address(rsp, 0), xmm15); //save last_key from xmm15
-      load_key(xmm15, key, 0xd0); // 0xd0; 256-bit key goes up to 0xe0
-      __ movdqu(Address(rsp, 6 * wordSize), xmm15);
-      load_key(xmm1, key, 0xe0);  // 0xe0;
-      __ movdqu(Address(rsp, 8 * wordSize), xmm1);
-      load_key(xmm15, key, 0xb0); // 0xb0;
-      __ movdqu(Address(rsp, 2 * wordSize), xmm15);
-      load_key(xmm1, key, 0xc0);  // 0xc0;
-      __ movdqu(Address(rsp, 4 * wordSize), xmm1);
-    }
-    __ align(OptoLoopAlignment);
-    __ BIND(L_multiBlock_loopTop[k]);
-    __ cmpptr(len_reg, PARALLEL_FACTOR * AESBlockSize); // see if at least 4 blocks left
-    __ jcc(Assembler::less, L_singleBlock_loopTopHead[k]);
-
-    if  (k != 0) {
-      __ movdqu(xmm15, Address(rsp, 2 * wordSize));
-      __ movdqu(xmm1, Address(rsp, 4 * wordSize));
-    }
-
-    __ movdqu(xmm_result0, Address(from, pos, Address::times_1, 0 * AESBlockSize)); // get next 4 blocks into xmmresult registers
-    __ movdqu(xmm_result1, Address(from, pos, Address::times_1, 1 * AESBlockSize));
-    __ movdqu(xmm_result2, Address(from, pos, Address::times_1, 2 * AESBlockSize));
-    __ movdqu(xmm_result3, Address(from, pos, Address::times_1, 3 * AESBlockSize));
-
-    DoFour(pxor, xmm_key_first);
-    if (k == 0) {
-      for (int rnum = 1; rnum < ROUNDS[k]; rnum++) {
-        DoFour(aesdec, as_XMMRegister(rnum + XMM_REG_NUM_KEY_FIRST));
-      }
-      DoFour(aesdeclast, xmm_key_last);
-    } else if (k == 1) {
-      for (int rnum = 1; rnum <= ROUNDS[k]-2; rnum++) {
-        DoFour(aesdec, as_XMMRegister(rnum + XMM_REG_NUM_KEY_FIRST));
-      }
-      __ movdqu(xmm_key_last, Address(rsp, 0)); // xmm15 needs to be loaded again.
-      DoFour(aesdec, xmm1);  // key : 0xc0
-      __ movdqu(xmm_prev_block_cipher, Address(rvec, 0x00));  // xmm1 needs to be loaded again
-      DoFour(aesdeclast, xmm_key_last);
-    } else if (k == 2) {
-      for (int rnum = 1; rnum <= ROUNDS[k] - 4; rnum++) {
-        DoFour(aesdec, as_XMMRegister(rnum + XMM_REG_NUM_KEY_FIRST));
-      }
-      DoFour(aesdec, xmm1);  // key : 0xc0
-      __ movdqu(xmm15, Address(rsp, 6 * wordSize));
-      __ movdqu(xmm1, Address(rsp, 8 * wordSize));
-      DoFour(aesdec, xmm15);  // key : 0xd0
-      __ movdqu(xmm_key_last, Address(rsp, 0)); // xmm15 needs to be loaded again.
-      DoFour(aesdec, xmm1);  // key : 0xe0
-      __ movdqu(xmm_prev_block_cipher, Address(rvec, 0x00));  // xmm1 needs to be loaded again
-      DoFour(aesdeclast, xmm_key_last);
-    }
-
-    // for each result, xor with the r vector of previous cipher block
-    __ pxor(xmm_result0, xmm_prev_block_cipher);
-    __ movdqu(xmm_prev_block_cipher, Address(from, pos, Address::times_1, 0 * AESBlockSize));
-    __ pxor(xmm_result1, xmm_prev_block_cipher);
-    __ movdqu(xmm_prev_block_cipher, Address(from, pos, Address::times_1, 1 * AESBlockSize));
-    __ pxor(xmm_result2, xmm_prev_block_cipher);
-    __ movdqu(xmm_prev_block_cipher, Address(from, pos, Address::times_1, 2 * AESBlockSize));
-    __ pxor(xmm_result3, xmm_prev_block_cipher);
-    __ movdqu(xmm_prev_block_cipher, Address(from, pos, Address::times_1, 3 * AESBlockSize));   // this will carry over to next set of blocks
-    if (k != 0) {
-      __ movdqu(Address(rvec, 0x00), xmm_prev_block_cipher);
-    }
-
-    __ movdqu(Address(to, pos, Address::times_1, 0 * AESBlockSize), xmm_result0);     // store 4 results into the next 64 bytes of output
-    __ movdqu(Address(to, pos, Address::times_1, 1 * AESBlockSize), xmm_result1);
-    __ movdqu(Address(to, pos, Address::times_1, 2 * AESBlockSize), xmm_result2);
-    __ movdqu(Address(to, pos, Address::times_1, 3 * AESBlockSize), xmm_result3);
-
-    __ addptr(pos, PARALLEL_FACTOR * AESBlockSize);
-    __ subptr(len_reg, PARALLEL_FACTOR * AESBlockSize);
-    __ jmp(L_multiBlock_loopTop[k]);
-
-    // registers used in the non-parallelized loops
-    // xmm register assignments for the loops below
-    const XMMRegister xmm_result = xmm0;
-    const XMMRegister xmm_prev_block_cipher_save = xmm2;
-    const XMMRegister xmm_key11 = xmm3;
-    const XMMRegister xmm_key12 = xmm4;
-    const XMMRegister key_tmp = xmm4;
-
-    __ BIND(L_singleBlock_loopTopHead[k]);
-    if (k == 1) {
-      __ addptr(rsp, 6 * wordSize);
-    } else if (k == 2) {
-      __ addptr(rsp, 10 * wordSize);
-    }
-    __ cmpptr(len_reg, 0); // any blocks left??
-    __ jcc(Assembler::equal, L_exit);
-    __ BIND(L_singleBlock_loopTopHead2[k]);
-    if (k == 1) {
-      load_key(xmm_key11, key, 0xb0); // 0xb0; 192-bit key goes up to 0xc0
-      load_key(xmm_key12, key, 0xc0); // 0xc0; 192-bit key goes up to 0xc0
-    }
-    if (k == 2) {
-      load_key(xmm_key11, key, 0xb0); // 0xb0; 256-bit key goes up to 0xe0
-    }
-    __ align(OptoLoopAlignment);
-    __ BIND(L_singleBlock_loopTop[k]);
-    __ movdqu(xmm_result, Address(from, pos, Address::times_1, 0)); // get next 16 bytes of cipher input
-    __ movdqa(xmm_prev_block_cipher_save, xmm_result); // save for next r vector
-    __ pxor(xmm_result, xmm_key_first); // do the aes dec rounds
-    for (int rnum = 1; rnum <= 9 ; rnum++) {
-        __ aesdec(xmm_result, as_XMMRegister(rnum + XMM_REG_NUM_KEY_FIRST));
-    }
-    if (k == 1) {
-      __ aesdec(xmm_result, xmm_key11);
-      __ aesdec(xmm_result, xmm_key12);
-    }
-    if (k == 2) {
-      __ aesdec(xmm_result, xmm_key11);
-      load_key(key_tmp, key, 0xc0);
-      __ aesdec(xmm_result, key_tmp);
-      load_key(key_tmp, key, 0xd0);
-      __ aesdec(xmm_result, key_tmp);
-      load_key(key_tmp, key, 0xe0);
-      __ aesdec(xmm_result, key_tmp);
-    }
-
-    __ aesdeclast(xmm_result, xmm_key_last); // xmm15 always came from key+0
-    __ pxor(xmm_result, xmm_prev_block_cipher); // xor with the current r vector
-    __ movdqu(Address(to, pos, Address::times_1, 0), xmm_result); // store into the next 16 bytes of output
-    // no need to store r to memory until we exit
-    __ movdqa(xmm_prev_block_cipher, xmm_prev_block_cipher_save); // set up next r vector with cipher input from this block
-    __ addptr(pos, AESBlockSize);
-    __ subptr(len_reg, AESBlockSize);
-    __ jcc(Assembler::notEqual, L_singleBlock_loopTop[k]);
-    if (k != 2) {
-      __ jmp(L_exit);
-    }
-  } //for 128/192/256
-
-  __ BIND(L_exit);
-  __ movdqu(Address(rvec, 0), xmm_prev_block_cipher);     // final value of r stored in rvec of CipherBlockChaining object
-  __ pop(rbx);
-#ifdef _WIN64
-  __ movl(rax, len_mem);
-#else
-  __ pop(rax); // return length
-#endif
-  __ leave(); // required for proper stackwalking of RuntimeStub frame
-  __ ret(0);
-
-  return start;
-}
-
-address StubGenerator::generate_electronicCodeBook_encryptAESCrypt() {
-  __ align(CodeEntryAlignment);
-  StubCodeMark mark(this, "StubRoutines", "electronicCodeBook_encryptAESCrypt");
-  address start = __ pc();
-
-  const Register from = c_rarg0;  // source array address
-  const Register to = c_rarg1;  // destination array address
-  const Register key = c_rarg2;  // key array address
-  const Register len = c_rarg3;  // src len (must be multiple of blocksize 16)
-  __ enter(); // required for proper stackwalking of RuntimeStub frame
-  __ aesecb_encrypt(from, to, key, len);
-  __ vzeroupper();
-  __ leave(); // required for proper stackwalking of RuntimeStub frame
-  __ ret(0);
-
-  return start;
- }
-
-address StubGenerator::generate_electronicCodeBook_decryptAESCrypt() {
-  __ align(CodeEntryAlignment);
-  StubCodeMark mark(this, "StubRoutines", "electronicCodeBook_decryptAESCrypt");
-  address start = __ pc();
-
-  const Register from = c_rarg0;  // source array address
-  const Register to = c_rarg1;  // destination array address
-  const Register key = c_rarg2;  // key array address
-  const Register len = c_rarg3;  // src len (must be multiple of blocksize 16)
-  __ enter(); // required for proper stackwalking of RuntimeStub frame
-  __ aesecb_decrypt(from, to, key, len);
-  __ vzeroupper();
-  __ leave(); // required for proper stackwalking of RuntimeStub frame
-  __ ret(0);
-
-  return start;
-}
-
 // ofs and limit are use for multi-block byte array.
 // int com.sun.security.provider.MD5.implCompress(byte[] b, int ofs)
 address StubGenerator::generate_md5_implCompress(bool multi_block, const char *name) {
@@ -2152,1051 +1475,6 @@ address StubGenerator::generate_sha512_implCompress(bool multi_block, const char
   buf, state, ofs, limit, rsp, multi_block, shuf_mask);
 
   __ vzeroupper();
-  __ leave();
-  __ ret(0);
-
-  return start;
-}
-
-address StubGenerator::ghash_polynomial512_addr() {
-  __ align(CodeEntryAlignment);
-  StubCodeMark mark(this, "StubRoutines", "_ghash_poly512_addr");
-  address start = __ pc();
-
-  __ emit_data64(0x00000001C2000000, relocInfo::none); // POLY for reduction
-  __ emit_data64(0xC200000000000000, relocInfo::none);
-  __ emit_data64(0x00000001C2000000, relocInfo::none);
-  __ emit_data64(0xC200000000000000, relocInfo::none);
-  __ emit_data64(0x00000001C2000000, relocInfo::none);
-  __ emit_data64(0xC200000000000000, relocInfo::none);
-  __ emit_data64(0x00000001C2000000, relocInfo::none);
-  __ emit_data64(0xC200000000000000, relocInfo::none);
-  __ emit_data64(0x0000000000000001, relocInfo::none); // POLY
-  __ emit_data64(0xC200000000000000, relocInfo::none);
-  __ emit_data64(0x0000000000000001, relocInfo::none); // TWOONE
-  __ emit_data64(0x0000000100000000, relocInfo::none);
-
-  return start;
-}
-
-// Vector AES Galois Counter Mode implementation.
-//
-// Inputs:           Windows    |   Linux
-//   in         = rcx (c_rarg0) | rsi (c_rarg0)
-//   len        = rdx (c_rarg1) | rdi (c_rarg1)
-//   ct         = r8  (c_rarg2) | rdx (c_rarg2)
-//   out        = r9  (c_rarg3) | rcx (c_rarg3)
-//   key        = r10           | r8  (c_rarg4)
-//   state      = r13           | r9  (c_rarg5)
-//   subkeyHtbl = r14           | r11
-//   counter    = rsi           | r12
-//
-// Output:
-//   rax - number of processed bytes
-address StubGenerator::generate_galoisCounterMode_AESCrypt() {
-  __ align(CodeEntryAlignment);
-  StubCodeMark mark(this, "StubRoutines", "galoisCounterMode_AESCrypt");
-  address start = __ pc();
-
-  const Register in = c_rarg0;
-  const Register len = c_rarg1;
-  const Register ct = c_rarg2;
-  const Register out = c_rarg3;
-  // and updated with the incremented counter in the end
-#ifndef _WIN64
-  const Register key = c_rarg4;
-  const Register state = c_rarg5;
-  const Address subkeyH_mem(rbp, 2 * wordSize);
-  const Register subkeyHtbl = r11;
-  const Register avx512_subkeyHtbl = r13;
-  const Address counter_mem(rbp, 3 * wordSize);
-  const Register counter = r12;
-#else
-  const Address key_mem(rbp, 6 * wordSize);
-  const Register key = r10;
-  const Address state_mem(rbp, 7 * wordSize);
-  const Register state = r13;
-  const Address subkeyH_mem(rbp, 8 * wordSize);
-  const Register subkeyHtbl = r14;
-  const Register avx512_subkeyHtbl = r12;
-  const Address counter_mem(rbp, 9 * wordSize);
-  const Register counter = rsi;
-#endif
-  __ enter();
- // Save state before entering routine
-  __ push(r12);
-  __ push(r13);
-  __ push(r14);
-  __ push(r15);
-  __ push(rbx);
-#ifdef _WIN64
-  // on win64, fill len_reg from stack position
-  __ push(rsi);
-  __ movptr(key, key_mem);
-  __ movptr(state, state_mem);
-#endif
-  __ movptr(subkeyHtbl, subkeyH_mem);
-  __ movptr(counter, counter_mem);
-// Save rbp and rsp
-  __ push(rbp);
-  __ movq(rbp, rsp);
-// Align stack
-  __ andq(rsp, -64);
-  __ subptr(rsp, 96 * longSize); // Create space on the stack for htbl entries
-  __ movptr(avx512_subkeyHtbl, rsp);
-
-  __ aesgcm_encrypt(in, len, ct, out, key, state, subkeyHtbl, avx512_subkeyHtbl, counter);
-  __ vzeroupper();
-
-  __ movq(rsp, rbp);
-  __ pop(rbp);
-
-  // Restore state before leaving routine
-#ifdef _WIN64
-  __ pop(rsi);
-#endif
-  __ pop(rbx);
-  __ pop(r15);
-  __ pop(r14);
-  __ pop(r13);
-  __ pop(r12);
-
-  __ leave(); // required for proper stackwalking of RuntimeStub frame
-  __ ret(0);
-
-  return start;
-}
-
-// This mask is used for incrementing counter value(linc0, linc4, etc.)
-address StubGenerator::counter_mask_addr() {
-  __ align64();
-  StubCodeMark mark(this, "StubRoutines", "counter_mask_addr");
-  address start = __ pc();
-
-  __ emit_data64(0x08090a0b0c0d0e0f, relocInfo::none);//lbswapmask
-  __ emit_data64(0x0001020304050607, relocInfo::none);
-  __ emit_data64(0x08090a0b0c0d0e0f, relocInfo::none);
-  __ emit_data64(0x0001020304050607, relocInfo::none);
-  __ emit_data64(0x08090a0b0c0d0e0f, relocInfo::none);
-  __ emit_data64(0x0001020304050607, relocInfo::none);
-  __ emit_data64(0x08090a0b0c0d0e0f, relocInfo::none);
-  __ emit_data64(0x0001020304050607, relocInfo::none);
-  __ emit_data64(0x0000000000000000, relocInfo::none);//linc0 = counter_mask_addr+64
-  __ emit_data64(0x0000000000000000, relocInfo::none);
-  __ emit_data64(0x0000000000000001, relocInfo::none);//counter_mask_addr() + 80
-  __ emit_data64(0x0000000000000000, relocInfo::none);
-  __ emit_data64(0x0000000000000002, relocInfo::none);
-  __ emit_data64(0x0000000000000000, relocInfo::none);
-  __ emit_data64(0x0000000000000003, relocInfo::none);
-  __ emit_data64(0x0000000000000000, relocInfo::none);
-  __ emit_data64(0x0000000000000004, relocInfo::none);//linc4 = counter_mask_addr() + 128
-  __ emit_data64(0x0000000000000000, relocInfo::none);
-  __ emit_data64(0x0000000000000004, relocInfo::none);
-  __ emit_data64(0x0000000000000000, relocInfo::none);
-  __ emit_data64(0x0000000000000004, relocInfo::none);
-  __ emit_data64(0x0000000000000000, relocInfo::none);
-  __ emit_data64(0x0000000000000004, relocInfo::none);
-  __ emit_data64(0x0000000000000000, relocInfo::none);
-  __ emit_data64(0x0000000000000008, relocInfo::none);//linc8 = counter_mask_addr() + 192
-  __ emit_data64(0x0000000000000000, relocInfo::none);
-  __ emit_data64(0x0000000000000008, relocInfo::none);
-  __ emit_data64(0x0000000000000000, relocInfo::none);
-  __ emit_data64(0x0000000000000008, relocInfo::none);
-  __ emit_data64(0x0000000000000000, relocInfo::none);
-  __ emit_data64(0x0000000000000008, relocInfo::none);
-  __ emit_data64(0x0000000000000000, relocInfo::none);
-  __ emit_data64(0x0000000000000020, relocInfo::none);//linc32 = counter_mask_addr() + 256
-  __ emit_data64(0x0000000000000000, relocInfo::none);
-  __ emit_data64(0x0000000000000020, relocInfo::none);
-  __ emit_data64(0x0000000000000000, relocInfo::none);
-  __ emit_data64(0x0000000000000020, relocInfo::none);
-  __ emit_data64(0x0000000000000000, relocInfo::none);
-  __ emit_data64(0x0000000000000020, relocInfo::none);
-  __ emit_data64(0x0000000000000000, relocInfo::none);
-  __ emit_data64(0x0000000000000010, relocInfo::none);//linc16 = counter_mask_addr() + 320
-  __ emit_data64(0x0000000000000000, relocInfo::none);
-  __ emit_data64(0x0000000000000010, relocInfo::none);
-  __ emit_data64(0x0000000000000000, relocInfo::none);
-  __ emit_data64(0x0000000000000010, relocInfo::none);
-  __ emit_data64(0x0000000000000000, relocInfo::none);
-  __ emit_data64(0x0000000000000010, relocInfo::none);
-  __ emit_data64(0x0000000000000000, relocInfo::none);
-
-  return start;
-}
-
- // Vector AES Counter implementation
-address StubGenerator::generate_counterMode_VectorAESCrypt()  {
-  __ align(CodeEntryAlignment);
-  StubCodeMark mark(this, "StubRoutines", "counterMode_AESCrypt");
-  address start = __ pc();
-
-  const Register from = c_rarg0; // source array address
-  const Register to = c_rarg1; // destination array address
-  const Register key = c_rarg2; // key array address r8
-  const Register counter = c_rarg3; // counter byte array initialized from counter array address
-  // and updated with the incremented counter in the end
-#ifndef _WIN64
-  const Register len_reg = c_rarg4;
-  const Register saved_encCounter_start = c_rarg5;
-  const Register used_addr = r10;
-  const Address  used_mem(rbp, 2 * wordSize);
-  const Register used = r11;
-#else
-  const Address len_mem(rbp, 6 * wordSize); // length is on stack on Win64
-  const Address saved_encCounter_mem(rbp, 7 * wordSize); // saved encrypted counter is on stack on Win64
-  const Address used_mem(rbp, 8 * wordSize); // used length is on stack on Win64
-  const Register len_reg = r10; // pick the first volatile windows register
-  const Register saved_encCounter_start = r11;
-  const Register used_addr = r13;
-  const Register used = r14;
-#endif
-  __ enter();
- // Save state before entering routine
-  __ push(r12);
-  __ push(r13);
-  __ push(r14);
-  __ push(r15);
-#ifdef _WIN64
-  // on win64, fill len_reg from stack position
-  __ movl(len_reg, len_mem);
-  __ movptr(saved_encCounter_start, saved_encCounter_mem);
-  __ movptr(used_addr, used_mem);
-  __ movl(used, Address(used_addr, 0));
-#else
-  __ push(len_reg); // Save
-  __ movptr(used_addr, used_mem);
-  __ movl(used, Address(used_addr, 0));
-#endif
-  __ push(rbx);
-  __ aesctr_encrypt(from, to, key, counter, len_reg, used, used_addr, saved_encCounter_start);
-  __ vzeroupper();
-  // Restore state before leaving routine
-  __ pop(rbx);
-#ifdef _WIN64
-  __ movl(rax, len_mem); // return length
-#else
-  __ pop(rax); // return length
-#endif
-  __ pop(r15);
-  __ pop(r14);
-  __ pop(r13);
-  __ pop(r12);
-
-  __ leave(); // required for proper stackwalking of RuntimeStub frame
-  __ ret(0);
-
-  return start;
-}
-
-// This is a version of CTR/AES crypt which does 6 blocks in a loop at a time
-// to hide instruction latency
-//
-// Arguments:
-//
-// Inputs:
-//   c_rarg0   - source byte array address
-//   c_rarg1   - destination byte array address
-//   c_rarg2   - K (key) in little endian int array
-//   c_rarg3   - counter vector byte array address
-//   Linux
-//     c_rarg4   -          input length
-//     c_rarg5   -          saved encryptedCounter start
-//     rbp + 6 * wordSize - saved used length
-//   Windows
-//     rbp + 6 * wordSize - input length
-//     rbp + 7 * wordSize - saved encryptedCounter start
-//     rbp + 8 * wordSize - saved used length
-//
-// Output:
-//   rax       - input length
-//
-address StubGenerator::generate_counterMode_AESCrypt_Parallel() {
-  assert(UseAES, "need AES instructions and misaligned SSE support");
-  __ align(CodeEntryAlignment);
-  StubCodeMark mark(this, "StubRoutines", "counterMode_AESCrypt");
-  address start = __ pc();
-
-  const Register from = c_rarg0; // source array address
-  const Register to = c_rarg1; // destination array address
-  const Register key = c_rarg2; // key array address
-  const Register counter = c_rarg3; // counter byte array initialized from counter array address
-                                    // and updated with the incremented counter in the end
-#ifndef _WIN64
-  const Register len_reg = c_rarg4;
-  const Register saved_encCounter_start = c_rarg5;
-  const Register used_addr = r10;
-  const Address  used_mem(rbp, 2 * wordSize);
-  const Register used = r11;
-#else
-  const Address len_mem(rbp, 6 * wordSize); // length is on stack on Win64
-  const Address saved_encCounter_mem(rbp, 7 * wordSize); // length is on stack on Win64
-  const Address used_mem(rbp, 8 * wordSize); // length is on stack on Win64
-  const Register len_reg = r10; // pick the first volatile windows register
-  const Register saved_encCounter_start = r11;
-  const Register used_addr = r13;
-  const Register used = r14;
-#endif
-  const Register pos = rax;
-
-  const int PARALLEL_FACTOR = 6;
-  const XMMRegister xmm_counter_shuf_mask = xmm0;
-  const XMMRegister xmm_key_shuf_mask = xmm1; // used temporarily to swap key bytes up front
-  const XMMRegister xmm_curr_counter = xmm2;
-
-  const XMMRegister xmm_key_tmp0 = xmm3;
-  const XMMRegister xmm_key_tmp1 = xmm4;
-
-  // registers holding the four results in the parallelized loop
-  const XMMRegister xmm_result0 = xmm5;
-  const XMMRegister xmm_result1 = xmm6;
-  const XMMRegister xmm_result2 = xmm7;
-  const XMMRegister xmm_result3 = xmm8;
-  const XMMRegister xmm_result4 = xmm9;
-  const XMMRegister xmm_result5 = xmm10;
-
-  const XMMRegister xmm_from0 = xmm11;
-  const XMMRegister xmm_from1 = xmm12;
-  const XMMRegister xmm_from2 = xmm13;
-  const XMMRegister xmm_from3 = xmm14; //the last one is xmm14. we have to preserve it on WIN64.
-  const XMMRegister xmm_from4 = xmm3; //reuse xmm3~4. Because xmm_key_tmp0~1 are useless when loading input text
-  const XMMRegister xmm_from5 = xmm4;
-
-  //for key_128, key_192, key_256
-  const int rounds[3] = {10, 12, 14};
-  Label L_exit_preLoop, L_preLoop_start;
-  Label L_multiBlock_loopTop[3];
-  Label L_singleBlockLoopTop[3];
-  Label L__incCounter[3][6]; //for 6 blocks
-  Label L__incCounter_single[3]; //for single block, key128, key192, key256
-  Label L_processTail_insr[3], L_processTail_4_insr[3], L_processTail_2_insr[3], L_processTail_1_insr[3], L_processTail_exit_insr[3];
-  Label L_processTail_4_extr[3], L_processTail_2_extr[3], L_processTail_1_extr[3], L_processTail_exit_extr[3];
-
-  Label L_exit;
-
-  __ enter(); // required for proper stackwalking of RuntimeStub frame
-
-#ifdef _WIN64
-  // allocate spill slots for r13, r14
-  enum {
-      saved_r13_offset,
-      saved_r14_offset
-  };
-  __ subptr(rsp, 2 * wordSize);
-  __ movptr(Address(rsp, saved_r13_offset * wordSize), r13);
-  __ movptr(Address(rsp, saved_r14_offset * wordSize), r14);
-
-  // on win64, fill len_reg from stack position
-  __ movl(len_reg, len_mem);
-  __ movptr(saved_encCounter_start, saved_encCounter_mem);
-  __ movptr(used_addr, used_mem);
-  __ movl(used, Address(used_addr, 0));
-#else
-  __ push(len_reg); // Save
-  __ movptr(used_addr, used_mem);
-  __ movl(used, Address(used_addr, 0));
-#endif
-
-  __ push(rbx); // Save RBX
-  __ movdqu(xmm_curr_counter, Address(counter, 0x00)); // initialize counter with initial counter
-  __ movdqu(xmm_counter_shuf_mask, ExternalAddress(StubRoutines::x86::counter_shuffle_mask_addr()), pos); // pos as scratch
-  __ pshufb(xmm_curr_counter, xmm_counter_shuf_mask); //counter is shuffled
-  __ movptr(pos, 0);
-
-  // Use the partially used encrpyted counter from last invocation
-  __ BIND(L_preLoop_start);
-  __ cmpptr(used, 16);
-  __ jcc(Assembler::aboveEqual, L_exit_preLoop);
-    __ cmpptr(len_reg, 0);
-    __ jcc(Assembler::lessEqual, L_exit_preLoop);
-    __ movb(rbx, Address(saved_encCounter_start, used));
-    __ xorb(rbx, Address(from, pos));
-    __ movb(Address(to, pos), rbx);
-    __ addptr(pos, 1);
-    __ addptr(used, 1);
-    __ subptr(len_reg, 1);
-
-  __ jmp(L_preLoop_start);
-
-  __ BIND(L_exit_preLoop);
-  __ movl(Address(used_addr, 0), used);
-
-  // key length could be only {11, 13, 15} * 4 = {44, 52, 60}
-  __ movdqu(xmm_key_shuf_mask, ExternalAddress(StubRoutines::x86::key_shuffle_mask_addr()), rbx); // rbx as scratch
-  __ movl(rbx, Address(key, arrayOopDesc::length_offset_in_bytes() - arrayOopDesc::base_offset_in_bytes(T_INT)));
-  __ cmpl(rbx, 52);
-  __ jcc(Assembler::equal, L_multiBlock_loopTop[1]);
-  __ cmpl(rbx, 60);
-  __ jcc(Assembler::equal, L_multiBlock_loopTop[2]);
-
-#define CTR_DoSix(opc, src_reg)                \
-  __ opc(xmm_result0, src_reg);              \
-  __ opc(xmm_result1, src_reg);              \
-  __ opc(xmm_result2, src_reg);              \
-  __ opc(xmm_result3, src_reg);              \
-  __ opc(xmm_result4, src_reg);              \
-  __ opc(xmm_result5, src_reg);
-
-  // k == 0 :  generate code for key_128
-  // k == 1 :  generate code for key_192
-  // k == 2 :  generate code for key_256
-  for (int k = 0; k < 3; ++k) {
-    //multi blocks starts here
-    __ align(OptoLoopAlignment);
-    __ BIND(L_multiBlock_loopTop[k]);
-    __ cmpptr(len_reg, PARALLEL_FACTOR * AESBlockSize); // see if at least PARALLEL_FACTOR blocks left
-    __ jcc(Assembler::less, L_singleBlockLoopTop[k]);
-    load_key(xmm_key_tmp0, key, 0x00, xmm_key_shuf_mask);
-
-    //load, then increase counters
-    CTR_DoSix(movdqa, xmm_curr_counter);
-    inc_counter(rbx, xmm_result1, 0x01, L__incCounter[k][0]);
-    inc_counter(rbx, xmm_result2, 0x02, L__incCounter[k][1]);
-    inc_counter(rbx, xmm_result3, 0x03, L__incCounter[k][2]);
-    inc_counter(rbx, xmm_result4, 0x04, L__incCounter[k][3]);
-    inc_counter(rbx, xmm_result5,  0x05, L__incCounter[k][4]);
-    inc_counter(rbx, xmm_curr_counter, 0x06, L__incCounter[k][5]);
-    CTR_DoSix(pshufb, xmm_counter_shuf_mask); // after increased, shuffled counters back for PXOR
-    CTR_DoSix(pxor, xmm_key_tmp0);   //PXOR with Round 0 key
-
-    //load two ROUND_KEYs at a time
-    for (int i = 1; i < rounds[k]; ) {
-      load_key(xmm_key_tmp1, key, (0x10 * i), xmm_key_shuf_mask);
-      load_key(xmm_key_tmp0, key, (0x10 * (i+1)), xmm_key_shuf_mask);
-      CTR_DoSix(aesenc, xmm_key_tmp1);
-      i++;
-      if (i != rounds[k]) {
-        CTR_DoSix(aesenc, xmm_key_tmp0);
-      } else {
-        CTR_DoSix(aesenclast, xmm_key_tmp0);
-      }
-      i++;
-    }
-
-    // get next PARALLEL_FACTOR blocks into xmm_result registers
-    __ movdqu(xmm_from0, Address(from, pos, Address::times_1, 0 * AESBlockSize));
-    __ movdqu(xmm_from1, Address(from, pos, Address::times_1, 1 * AESBlockSize));
-    __ movdqu(xmm_from2, Address(from, pos, Address::times_1, 2 * AESBlockSize));
-    __ movdqu(xmm_from3, Address(from, pos, Address::times_1, 3 * AESBlockSize));
-    __ movdqu(xmm_from4, Address(from, pos, Address::times_1, 4 * AESBlockSize));
-    __ movdqu(xmm_from5, Address(from, pos, Address::times_1, 5 * AESBlockSize));
-
-    __ pxor(xmm_result0, xmm_from0);
-    __ pxor(xmm_result1, xmm_from1);
-    __ pxor(xmm_result2, xmm_from2);
-    __ pxor(xmm_result3, xmm_from3);
-    __ pxor(xmm_result4, xmm_from4);
-    __ pxor(xmm_result5, xmm_from5);
-
-    // store 6 results into the next 64 bytes of output
-    __ movdqu(Address(to, pos, Address::times_1, 0 * AESBlockSize), xmm_result0);
-    __ movdqu(Address(to, pos, Address::times_1, 1 * AESBlockSize), xmm_result1);
-    __ movdqu(Address(to, pos, Address::times_1, 2 * AESBlockSize), xmm_result2);
-    __ movdqu(Address(to, pos, Address::times_1, 3 * AESBlockSize), xmm_result3);
-    __ movdqu(Address(to, pos, Address::times_1, 4 * AESBlockSize), xmm_result4);
-    __ movdqu(Address(to, pos, Address::times_1, 5 * AESBlockSize), xmm_result5);
-
-    __ addptr(pos, PARALLEL_FACTOR * AESBlockSize); // increase the length of crypt text
-    __ subptr(len_reg, PARALLEL_FACTOR * AESBlockSize); // decrease the remaining length
-    __ jmp(L_multiBlock_loopTop[k]);
-
-    // singleBlock starts here
-    __ align(OptoLoopAlignment);
-    __ BIND(L_singleBlockLoopTop[k]);
-    __ cmpptr(len_reg, 0);
-    __ jcc(Assembler::lessEqual, L_exit);
-    load_key(xmm_key_tmp0, key, 0x00, xmm_key_shuf_mask);
-    __ movdqa(xmm_result0, xmm_curr_counter);
-    inc_counter(rbx, xmm_curr_counter, 0x01, L__incCounter_single[k]);
-    __ pshufb(xmm_result0, xmm_counter_shuf_mask);
-    __ pxor(xmm_result0, xmm_key_tmp0);
-    for (int i = 1; i < rounds[k]; i++) {
-      load_key(xmm_key_tmp0, key, (0x10 * i), xmm_key_shuf_mask);
-      __ aesenc(xmm_result0, xmm_key_tmp0);
-    }
-    load_key(xmm_key_tmp0, key, (rounds[k] * 0x10), xmm_key_shuf_mask);
-    __ aesenclast(xmm_result0, xmm_key_tmp0);
-    __ cmpptr(len_reg, AESBlockSize);
-    __ jcc(Assembler::less, L_processTail_insr[k]);
-      __ movdqu(xmm_from0, Address(from, pos, Address::times_1, 0 * AESBlockSize));
-      __ pxor(xmm_result0, xmm_from0);
-      __ movdqu(Address(to, pos, Address::times_1, 0 * AESBlockSize), xmm_result0);
-      __ addptr(pos, AESBlockSize);
-      __ subptr(len_reg, AESBlockSize);
-      __ jmp(L_singleBlockLoopTop[k]);
-    __ BIND(L_processTail_insr[k]);                               // Process the tail part of the input array
-      __ addptr(pos, len_reg);                                    // 1. Insert bytes from src array into xmm_from0 register
-      __ testptr(len_reg, 8);
-      __ jcc(Assembler::zero, L_processTail_4_insr[k]);
-        __ subptr(pos,8);
-        __ pinsrq(xmm_from0, Address(from, pos), 0);
-      __ BIND(L_processTail_4_insr[k]);
-      __ testptr(len_reg, 4);
-      __ jcc(Assembler::zero, L_processTail_2_insr[k]);
-        __ subptr(pos,4);
-        __ pslldq(xmm_from0, 4);
-        __ pinsrd(xmm_from0, Address(from, pos), 0);
-      __ BIND(L_processTail_2_insr[k]);
-      __ testptr(len_reg, 2);
-      __ jcc(Assembler::zero, L_processTail_1_insr[k]);
-        __ subptr(pos, 2);
-        __ pslldq(xmm_from0, 2);
-        __ pinsrw(xmm_from0, Address(from, pos), 0);
-      __ BIND(L_processTail_1_insr[k]);
-      __ testptr(len_reg, 1);
-      __ jcc(Assembler::zero, L_processTail_exit_insr[k]);
-        __ subptr(pos, 1);
-        __ pslldq(xmm_from0, 1);
-        __ pinsrb(xmm_from0, Address(from, pos), 0);
-      __ BIND(L_processTail_exit_insr[k]);
-
-      __ movdqu(Address(saved_encCounter_start, 0), xmm_result0);  // 2. Perform pxor of the encrypted counter and plaintext Bytes.
-      __ pxor(xmm_result0, xmm_from0);                             //    Also the encrypted counter is saved for next invocation.
-
-      __ testptr(len_reg, 8);
-      __ jcc(Assembler::zero, L_processTail_4_extr[k]);            // 3. Extract bytes from xmm_result0 into the dest. array
-        __ pextrq(Address(to, pos), xmm_result0, 0);
-        __ psrldq(xmm_result0, 8);
-        __ addptr(pos, 8);
-      __ BIND(L_processTail_4_extr[k]);
-      __ testptr(len_reg, 4);
-      __ jcc(Assembler::zero, L_processTail_2_extr[k]);
-        __ pextrd(Address(to, pos), xmm_result0, 0);
-        __ psrldq(xmm_result0, 4);
-        __ addptr(pos, 4);
-      __ BIND(L_processTail_2_extr[k]);
-      __ testptr(len_reg, 2);
-      __ jcc(Assembler::zero, L_processTail_1_extr[k]);
-        __ pextrw(Address(to, pos), xmm_result0, 0);
-        __ psrldq(xmm_result0, 2);
-        __ addptr(pos, 2);
-      __ BIND(L_processTail_1_extr[k]);
-      __ testptr(len_reg, 1);
-      __ jcc(Assembler::zero, L_processTail_exit_extr[k]);
-        __ pextrb(Address(to, pos), xmm_result0, 0);
-
-      __ BIND(L_processTail_exit_extr[k]);
-      __ movl(Address(used_addr, 0), len_reg);
-      __ jmp(L_exit);
-  }
-
-  __ BIND(L_exit);
-  __ pshufb(xmm_curr_counter, xmm_counter_shuf_mask); //counter is shuffled back.
-  __ movdqu(Address(counter, 0), xmm_curr_counter); //save counter back
-  __ pop(rbx); // pop the saved RBX.
-#ifdef _WIN64
-  __ movl(rax, len_mem);
-  __ movptr(r13, Address(rsp, saved_r13_offset * wordSize));
-  __ movptr(r14, Address(rsp, saved_r14_offset * wordSize));
-  __ addptr(rsp, 2 * wordSize);
-#else
-  __ pop(rax); // return 'len'
-#endif
-  __ leave(); // required for proper stackwalking of RuntimeStub frame
-  __ ret(0);
-
-  return start;
-}
-
-void StubGenerator::roundDec(XMMRegister xmm_reg) {
-  __ vaesdec(xmm1, xmm1, xmm_reg, Assembler::AVX_512bit);
-  __ vaesdec(xmm2, xmm2, xmm_reg, Assembler::AVX_512bit);
-  __ vaesdec(xmm3, xmm3, xmm_reg, Assembler::AVX_512bit);
-  __ vaesdec(xmm4, xmm4, xmm_reg, Assembler::AVX_512bit);
-  __ vaesdec(xmm5, xmm5, xmm_reg, Assembler::AVX_512bit);
-  __ vaesdec(xmm6, xmm6, xmm_reg, Assembler::AVX_512bit);
-  __ vaesdec(xmm7, xmm7, xmm_reg, Assembler::AVX_512bit);
-  __ vaesdec(xmm8, xmm8, xmm_reg, Assembler::AVX_512bit);
-}
-
-void StubGenerator::roundDeclast(XMMRegister xmm_reg) {
-  __ vaesdeclast(xmm1, xmm1, xmm_reg, Assembler::AVX_512bit);
-  __ vaesdeclast(xmm2, xmm2, xmm_reg, Assembler::AVX_512bit);
-  __ vaesdeclast(xmm3, xmm3, xmm_reg, Assembler::AVX_512bit);
-  __ vaesdeclast(xmm4, xmm4, xmm_reg, Assembler::AVX_512bit);
-  __ vaesdeclast(xmm5, xmm5, xmm_reg, Assembler::AVX_512bit);
-  __ vaesdeclast(xmm6, xmm6, xmm_reg, Assembler::AVX_512bit);
-  __ vaesdeclast(xmm7, xmm7, xmm_reg, Assembler::AVX_512bit);
-  __ vaesdeclast(xmm8, xmm8, xmm_reg, Assembler::AVX_512bit);
-}
-
-void StubGenerator::ev_load_key(XMMRegister xmmdst, Register key, int offset, XMMRegister xmm_shuf_mask) {
-  __ movdqu(xmmdst, Address(key, offset));
-  if (xmm_shuf_mask != xnoreg) {
-    __ pshufb(xmmdst, xmm_shuf_mask);
-  } else {
-    __ pshufb(xmmdst, ExternalAddress(StubRoutines::x86::key_shuffle_mask_addr()));
-  }
-  __ evshufi64x2(xmmdst, xmmdst, xmmdst, 0x0, Assembler::AVX_512bit);
-}
-
-address StubGenerator::generate_cipherBlockChaining_decryptVectorAESCrypt() {
-  assert(VM_Version::supports_avx512_vaes(), "need AES instructions and misaligned SSE support");
-  __ align(CodeEntryAlignment);
-  StubCodeMark mark(this, "StubRoutines", "cipherBlockChaining_decryptAESCrypt");
-  address start = __ pc();
-
-  const Register from = c_rarg0;  // source array address
-  const Register to = c_rarg1;  // destination array address
-  const Register key = c_rarg2;  // key array address
-  const Register rvec = c_rarg3;  // r byte array initialized from initvector array address
-  // and left with the results of the last encryption block
-#ifndef _WIN64
-  const Register len_reg = c_rarg4;  // src len (must be multiple of blocksize 16)
-#else
-  const Address  len_mem(rbp, 6 * wordSize);  // length is on stack on Win64
-  const Register len_reg = r11;      // pick the volatile windows register
-#endif
-
-  Label Loop, Loop1, L_128, L_256, L_192, KEY_192, KEY_256, Loop2, Lcbc_dec_rem_loop,
-        Lcbc_dec_rem_last, Lcbc_dec_ret, Lcbc_dec_rem, Lcbc_exit;
-
-  __ enter();
-
-#ifdef _WIN64
-// on win64, fill len_reg from stack position
-  __ movl(len_reg, len_mem);
-#else
-  __ push(len_reg); // Save
-#endif
-  __ push(rbx);
-  __ vzeroupper();
-
-  // Temporary variable declaration for swapping key bytes
-  const XMMRegister xmm_key_shuf_mask = xmm1;
-  __ movdqu(xmm_key_shuf_mask, ExternalAddress(StubRoutines::x86::key_shuffle_mask_addr()));
-
-  // Calculate number of rounds from key size: 44 for 10-rounds, 52 for 12-rounds, 60 for 14-rounds
-  const Register rounds = rbx;
-  __ movl(rounds, Address(key, arrayOopDesc::length_offset_in_bytes() - arrayOopDesc::base_offset_in_bytes(T_INT)));
-
-  const XMMRegister IV = xmm0;
-  // Load IV and broadcast value to 512-bits
-  __ evbroadcasti64x2(IV, Address(rvec, 0), Assembler::AVX_512bit);
-
-  // Temporary variables for storing round keys
-  const XMMRegister RK0 = xmm30;
-  const XMMRegister RK1 = xmm9;
-  const XMMRegister RK2 = xmm18;
-  const XMMRegister RK3 = xmm19;
-  const XMMRegister RK4 = xmm20;
-  const XMMRegister RK5 = xmm21;
-  const XMMRegister RK6 = xmm22;
-  const XMMRegister RK7 = xmm23;
-  const XMMRegister RK8 = xmm24;
-  const XMMRegister RK9 = xmm25;
-  const XMMRegister RK10 = xmm26;
-
-  // Load and shuffle key
-  // the java expanded key ordering is rotated one position from what we want
-  // so we start from 1*16 here and hit 0*16 last
-  ev_load_key(RK1, key, 1 * 16, xmm_key_shuf_mask);
-  ev_load_key(RK2, key, 2 * 16, xmm_key_shuf_mask);
-  ev_load_key(RK3, key, 3 * 16, xmm_key_shuf_mask);
-  ev_load_key(RK4, key, 4 * 16, xmm_key_shuf_mask);
-  ev_load_key(RK5, key, 5 * 16, xmm_key_shuf_mask);
-  ev_load_key(RK6, key, 6 * 16, xmm_key_shuf_mask);
-  ev_load_key(RK7, key, 7 * 16, xmm_key_shuf_mask);
-  ev_load_key(RK8, key, 8 * 16, xmm_key_shuf_mask);
-  ev_load_key(RK9, key, 9 * 16, xmm_key_shuf_mask);
-  ev_load_key(RK10, key, 10 * 16, xmm_key_shuf_mask);
-  ev_load_key(RK0, key, 0*16, xmm_key_shuf_mask);
-
-  // Variables for storing source cipher text
-  const XMMRegister S0 = xmm10;
-  const XMMRegister S1 = xmm11;
-  const XMMRegister S2 = xmm12;
-  const XMMRegister S3 = xmm13;
-  const XMMRegister S4 = xmm14;
-  const XMMRegister S5 = xmm15;
-  const XMMRegister S6 = xmm16;
-  const XMMRegister S7 = xmm17;
-
-  // Variables for storing decrypted text
-  const XMMRegister B0 = xmm1;
-  const XMMRegister B1 = xmm2;
-  const XMMRegister B2 = xmm3;
-  const XMMRegister B3 = xmm4;
-  const XMMRegister B4 = xmm5;
-  const XMMRegister B5 = xmm6;
-  const XMMRegister B6 = xmm7;
-  const XMMRegister B7 = xmm8;
-
-  __ cmpl(rounds, 44);
-  __ jcc(Assembler::greater, KEY_192);
-  __ jmp(Loop);
-
-  __ BIND(KEY_192);
-  const XMMRegister RK11 = xmm27;
-  const XMMRegister RK12 = xmm28;
-  ev_load_key(RK11, key, 11*16, xmm_key_shuf_mask);
-  ev_load_key(RK12, key, 12*16, xmm_key_shuf_mask);
-
-  __ cmpl(rounds, 52);
-  __ jcc(Assembler::greater, KEY_256);
-  __ jmp(Loop);
-
-  __ BIND(KEY_256);
-  const XMMRegister RK13 = xmm29;
-  const XMMRegister RK14 = xmm31;
-  ev_load_key(RK13, key, 13*16, xmm_key_shuf_mask);
-  ev_load_key(RK14, key, 14*16, xmm_key_shuf_mask);
-
-  __ BIND(Loop);
-  __ cmpl(len_reg, 512);
-  __ jcc(Assembler::below, Lcbc_dec_rem);
-  __ BIND(Loop1);
-  __ subl(len_reg, 512);
-  __ evmovdquq(S0, Address(from, 0 * 64), Assembler::AVX_512bit);
-  __ evmovdquq(S1, Address(from, 1 * 64), Assembler::AVX_512bit);
-  __ evmovdquq(S2, Address(from, 2 * 64), Assembler::AVX_512bit);
-  __ evmovdquq(S3, Address(from, 3 * 64), Assembler::AVX_512bit);
-  __ evmovdquq(S4, Address(from, 4 * 64), Assembler::AVX_512bit);
-  __ evmovdquq(S5, Address(from, 5 * 64), Assembler::AVX_512bit);
-  __ evmovdquq(S6, Address(from, 6 * 64), Assembler::AVX_512bit);
-  __ evmovdquq(S7, Address(from, 7 * 64), Assembler::AVX_512bit);
-  __ leaq(from, Address(from, 8 * 64));
-
-  __ evpxorq(B0, S0, RK1, Assembler::AVX_512bit);
-  __ evpxorq(B1, S1, RK1, Assembler::AVX_512bit);
-  __ evpxorq(B2, S2, RK1, Assembler::AVX_512bit);
-  __ evpxorq(B3, S3, RK1, Assembler::AVX_512bit);
-  __ evpxorq(B4, S4, RK1, Assembler::AVX_512bit);
-  __ evpxorq(B5, S5, RK1, Assembler::AVX_512bit);
-  __ evpxorq(B6, S6, RK1, Assembler::AVX_512bit);
-  __ evpxorq(B7, S7, RK1, Assembler::AVX_512bit);
-
-  __ evalignq(IV, S0, IV, 0x06);
-  __ evalignq(S0, S1, S0, 0x06);
-  __ evalignq(S1, S2, S1, 0x06);
-  __ evalignq(S2, S3, S2, 0x06);
-  __ evalignq(S3, S4, S3, 0x06);
-  __ evalignq(S4, S5, S4, 0x06);
-  __ evalignq(S5, S6, S5, 0x06);
-  __ evalignq(S6, S7, S6, 0x06);
-
-  roundDec(RK2);
-  roundDec(RK3);
-  roundDec(RK4);
-  roundDec(RK5);
-  roundDec(RK6);
-  roundDec(RK7);
-  roundDec(RK8);
-  roundDec(RK9);
-  roundDec(RK10);
-
-  __ cmpl(rounds, 44);
-  __ jcc(Assembler::belowEqual, L_128);
-  roundDec(RK11);
-  roundDec(RK12);
-
-  __ cmpl(rounds, 52);
-  __ jcc(Assembler::belowEqual, L_192);
-  roundDec(RK13);
-  roundDec(RK14);
-
-  __ BIND(L_256);
-  roundDeclast(RK0);
-  __ jmp(Loop2);
-
-  __ BIND(L_128);
-  roundDeclast(RK0);
-  __ jmp(Loop2);
-
-  __ BIND(L_192);
-  roundDeclast(RK0);
-
-  __ BIND(Loop2);
-  __ evpxorq(B0, B0, IV, Assembler::AVX_512bit);
-  __ evpxorq(B1, B1, S0, Assembler::AVX_512bit);
-  __ evpxorq(B2, B2, S1, Assembler::AVX_512bit);
-  __ evpxorq(B3, B3, S2, Assembler::AVX_512bit);
-  __ evpxorq(B4, B4, S3, Assembler::AVX_512bit);
-  __ evpxorq(B5, B5, S4, Assembler::AVX_512bit);
-  __ evpxorq(B6, B6, S5, Assembler::AVX_512bit);
-  __ evpxorq(B7, B7, S6, Assembler::AVX_512bit);
-  __ evmovdquq(IV, S7, Assembler::AVX_512bit);
-
-  __ evmovdquq(Address(to, 0 * 64), B0, Assembler::AVX_512bit);
-  __ evmovdquq(Address(to, 1 * 64), B1, Assembler::AVX_512bit);
-  __ evmovdquq(Address(to, 2 * 64), B2, Assembler::AVX_512bit);
-  __ evmovdquq(Address(to, 3 * 64), B3, Assembler::AVX_512bit);
-  __ evmovdquq(Address(to, 4 * 64), B4, Assembler::AVX_512bit);
-  __ evmovdquq(Address(to, 5 * 64), B5, Assembler::AVX_512bit);
-  __ evmovdquq(Address(to, 6 * 64), B6, Assembler::AVX_512bit);
-  __ evmovdquq(Address(to, 7 * 64), B7, Assembler::AVX_512bit);
-  __ leaq(to, Address(to, 8 * 64));
-  __ jmp(Loop);
-
-  __ BIND(Lcbc_dec_rem);
-  __ evshufi64x2(IV, IV, IV, 0x03, Assembler::AVX_512bit);
-
-  __ BIND(Lcbc_dec_rem_loop);
-  __ subl(len_reg, 16);
-  __ jcc(Assembler::carrySet, Lcbc_dec_ret);
-
-  __ movdqu(S0, Address(from, 0));
-  __ evpxorq(B0, S0, RK1, Assembler::AVX_512bit);
-  __ vaesdec(B0, B0, RK2, Assembler::AVX_512bit);
-  __ vaesdec(B0, B0, RK3, Assembler::AVX_512bit);
-  __ vaesdec(B0, B0, RK4, Assembler::AVX_512bit);
-  __ vaesdec(B0, B0, RK5, Assembler::AVX_512bit);
-  __ vaesdec(B0, B0, RK6, Assembler::AVX_512bit);
-  __ vaesdec(B0, B0, RK7, Assembler::AVX_512bit);
-  __ vaesdec(B0, B0, RK8, Assembler::AVX_512bit);
-  __ vaesdec(B0, B0, RK9, Assembler::AVX_512bit);
-  __ vaesdec(B0, B0, RK10, Assembler::AVX_512bit);
-  __ cmpl(rounds, 44);
-  __ jcc(Assembler::belowEqual, Lcbc_dec_rem_last);
-
-  __ vaesdec(B0, B0, RK11, Assembler::AVX_512bit);
-  __ vaesdec(B0, B0, RK12, Assembler::AVX_512bit);
-  __ cmpl(rounds, 52);
-  __ jcc(Assembler::belowEqual, Lcbc_dec_rem_last);
-
-  __ vaesdec(B0, B0, RK13, Assembler::AVX_512bit);
-  __ vaesdec(B0, B0, RK14, Assembler::AVX_512bit);
-
-  __ BIND(Lcbc_dec_rem_last);
-  __ vaesdeclast(B0, B0, RK0, Assembler::AVX_512bit);
-
-  __ evpxorq(B0, B0, IV, Assembler::AVX_512bit);
-  __ evmovdquq(IV, S0, Assembler::AVX_512bit);
-  __ movdqu(Address(to, 0), B0);
-  __ leaq(from, Address(from, 16));
-  __ leaq(to, Address(to, 16));
-  __ jmp(Lcbc_dec_rem_loop);
-
-  __ BIND(Lcbc_dec_ret);
-  __ movdqu(Address(rvec, 0), IV);
-
-  // Zero out the round keys
-  __ evpxorq(RK0, RK0, RK0, Assembler::AVX_512bit);
-  __ evpxorq(RK1, RK1, RK1, Assembler::AVX_512bit);
-  __ evpxorq(RK2, RK2, RK2, Assembler::AVX_512bit);
-  __ evpxorq(RK3, RK3, RK3, Assembler::AVX_512bit);
-  __ evpxorq(RK4, RK4, RK4, Assembler::AVX_512bit);
-  __ evpxorq(RK5, RK5, RK5, Assembler::AVX_512bit);
-  __ evpxorq(RK6, RK6, RK6, Assembler::AVX_512bit);
-  __ evpxorq(RK7, RK7, RK7, Assembler::AVX_512bit);
-  __ evpxorq(RK8, RK8, RK8, Assembler::AVX_512bit);
-  __ evpxorq(RK9, RK9, RK9, Assembler::AVX_512bit);
-  __ evpxorq(RK10, RK10, RK10, Assembler::AVX_512bit);
-  __ cmpl(rounds, 44);
-  __ jcc(Assembler::belowEqual, Lcbc_exit);
-  __ evpxorq(RK11, RK11, RK11, Assembler::AVX_512bit);
-  __ evpxorq(RK12, RK12, RK12, Assembler::AVX_512bit);
-  __ cmpl(rounds, 52);
-  __ jcc(Assembler::belowEqual, Lcbc_exit);
-  __ evpxorq(RK13, RK13, RK13, Assembler::AVX_512bit);
-  __ evpxorq(RK14, RK14, RK14, Assembler::AVX_512bit);
-
-  __ BIND(Lcbc_exit);
-  __ vzeroupper();
-  __ pop(rbx);
-#ifdef _WIN64
-  __ movl(rax, len_mem);
-#else
-  __ pop(rax); // return length
-#endif
-  __ leave(); // required for proper stackwalking of RuntimeStub frame
-  __ ret(0);
-
-  return start;
-}
-
-// Polynomial x^128+x^127+x^126+x^121+1
-address StubGenerator::ghash_polynomial_addr() {
-  __ align(CodeEntryAlignment);
-  StubCodeMark mark(this, "StubRoutines", "_ghash_poly_addr");
-  address start = __ pc();
-
-  __ emit_data64(0x0000000000000001, relocInfo::none);
-  __ emit_data64(0xc200000000000000, relocInfo::none);
-
-  return start;
-}
-
-address StubGenerator::ghash_shufflemask_addr() {
-  __ align(CodeEntryAlignment);
-  StubCodeMark mark(this, "StubRoutines", "_ghash_shuffmask_addr");
-  address start = __ pc();
-
-  __ emit_data64(0x0f0f0f0f0f0f0f0f, relocInfo::none);
-  __ emit_data64(0x0f0f0f0f0f0f0f0f, relocInfo::none);
-
-  return start;
-}
-
-// Ghash single and multi block operations using AVX instructions
-address StubGenerator::generate_avx_ghash_processBlocks() {
-  __ align(CodeEntryAlignment);
-
-  StubCodeMark mark(this, "StubRoutines", "ghash_processBlocks");
-  address start = __ pc();
-
-  // arguments
-  const Register state = c_rarg0;
-  const Register htbl = c_rarg1;
-  const Register data = c_rarg2;
-  const Register blocks = c_rarg3;
-  __ enter();
- // Save state before entering routine
-  __ avx_ghash(state, htbl, data, blocks);
-  __ leave(); // required for proper stackwalking of RuntimeStub frame
-  __ ret(0);
-
-  return start;
-}
-
-// byte swap x86 long
-address StubGenerator::generate_ghash_long_swap_mask() {
-  __ align(CodeEntryAlignment);
-  StubCodeMark mark(this, "StubRoutines", "ghash_long_swap_mask");
-  address start = __ pc();
-
-  __ emit_data64(0x0f0e0d0c0b0a0908, relocInfo::none );
-  __ emit_data64(0x0706050403020100, relocInfo::none );
-
-return start;
-}
-
-// byte swap x86 byte array
-address StubGenerator::generate_ghash_byte_swap_mask() {
-  __ align(CodeEntryAlignment);
-  StubCodeMark mark(this, "StubRoutines", "ghash_byte_swap_mask");
-  address start = __ pc();
-
-  __ emit_data64(0x08090a0b0c0d0e0f, relocInfo::none );
-  __ emit_data64(0x0001020304050607, relocInfo::none );
-
-return start;
-}
-
-/* Single and multi-block ghash operations */
-address StubGenerator::generate_ghash_processBlocks() {
-  __ align(CodeEntryAlignment);
-  Label L_ghash_loop, L_exit;
-  StubCodeMark mark(this, "StubRoutines", "ghash_processBlocks");
-  address start = __ pc();
-
-  const Register state        = c_rarg0;
-  const Register subkeyH      = c_rarg1;
-  const Register data         = c_rarg2;
-  const Register blocks       = c_rarg3;
-
-  const XMMRegister xmm_temp0 = xmm0;
-  const XMMRegister xmm_temp1 = xmm1;
-  const XMMRegister xmm_temp2 = xmm2;
-  const XMMRegister xmm_temp3 = xmm3;
-  const XMMRegister xmm_temp4 = xmm4;
-  const XMMRegister xmm_temp5 = xmm5;
-  const XMMRegister xmm_temp6 = xmm6;
-  const XMMRegister xmm_temp7 = xmm7;
-  const XMMRegister xmm_temp8 = xmm8;
-  const XMMRegister xmm_temp9 = xmm9;
-  const XMMRegister xmm_temp10 = xmm10;
-
-  __ enter();
-
-  __ movdqu(xmm_temp10, ExternalAddress(StubRoutines::x86::ghash_long_swap_mask_addr()));
-
-  __ movdqu(xmm_temp0, Address(state, 0));
-  __ pshufb(xmm_temp0, xmm_temp10);
-
-
-  __ BIND(L_ghash_loop);
-  __ movdqu(xmm_temp2, Address(data, 0));
-  __ pshufb(xmm_temp2, ExternalAddress(StubRoutines::x86::ghash_byte_swap_mask_addr()));
-
-  __ movdqu(xmm_temp1, Address(subkeyH, 0));
-  __ pshufb(xmm_temp1, xmm_temp10);
-
-  __ pxor(xmm_temp0, xmm_temp2);
-
-  //
-  // Multiply with the hash key
-  //
-  __ movdqu(xmm_temp3, xmm_temp0);
-  __ pclmulqdq(xmm_temp3, xmm_temp1, 0);      // xmm3 holds a0*b0
-  __ movdqu(xmm_temp4, xmm_temp0);
-  __ pclmulqdq(xmm_temp4, xmm_temp1, 16);     // xmm4 holds a0*b1
-
-  __ movdqu(xmm_temp5, xmm_temp0);
-  __ pclmulqdq(xmm_temp5, xmm_temp1, 1);      // xmm5 holds a1*b0
-  __ movdqu(xmm_temp6, xmm_temp0);
-  __ pclmulqdq(xmm_temp6, xmm_temp1, 17);     // xmm6 holds a1*b1
-
-  __ pxor(xmm_temp4, xmm_temp5);      // xmm4 holds a0*b1 + a1*b0
-
-  __ movdqu(xmm_temp5, xmm_temp4);    // move the contents of xmm4 to xmm5
-  __ psrldq(xmm_temp4, 8);    // shift by xmm4 64 bits to the right
-  __ pslldq(xmm_temp5, 8);    // shift by xmm5 64 bits to the left
-  __ pxor(xmm_temp3, xmm_temp5);
-  __ pxor(xmm_temp6, xmm_temp4);      // Register pair <xmm6:xmm3> holds the result
-                                      // of the carry-less multiplication of
-                                      // xmm0 by xmm1.
-
-  // We shift the result of the multiplication by one bit position
-  // to the left to cope for the fact that the bits are reversed.
-  __ movdqu(xmm_temp7, xmm_temp3);
-  __ movdqu(xmm_temp8, xmm_temp6);
-  __ pslld(xmm_temp3, 1);
-  __ pslld(xmm_temp6, 1);
-  __ psrld(xmm_temp7, 31);
-  __ psrld(xmm_temp8, 31);
-  __ movdqu(xmm_temp9, xmm_temp7);
-  __ pslldq(xmm_temp8, 4);
-  __ pslldq(xmm_temp7, 4);
-  __ psrldq(xmm_temp9, 12);
-  __ por(xmm_temp3, xmm_temp7);
-  __ por(xmm_temp6, xmm_temp8);
-  __ por(xmm_temp6, xmm_temp9);
-
-  //
-  // First phase of the reduction
-  //
-  // Move xmm3 into xmm7, xmm8, xmm9 in order to perform the shifts
-  // independently.
-  __ movdqu(xmm_temp7, xmm_temp3);
-  __ movdqu(xmm_temp8, xmm_temp3);
-  __ movdqu(xmm_temp9, xmm_temp3);
-  __ pslld(xmm_temp7, 31);    // packed right shift shifting << 31
-  __ pslld(xmm_temp8, 30);    // packed right shift shifting << 30
-  __ pslld(xmm_temp9, 25);    // packed right shift shifting << 25
-  __ pxor(xmm_temp7, xmm_temp8);      // xor the shifted versions
-  __ pxor(xmm_temp7, xmm_temp9);
-  __ movdqu(xmm_temp8, xmm_temp7);
-  __ pslldq(xmm_temp7, 12);
-  __ psrldq(xmm_temp8, 4);
-  __ pxor(xmm_temp3, xmm_temp7);      // first phase of the reduction complete
-
-  //
-  // Second phase of the reduction
-  //
-  // Make 3 copies of xmm3 in xmm2, xmm4, xmm5 for doing these
-  // shift operations.
-  __ movdqu(xmm_temp2, xmm_temp3);
-  __ movdqu(xmm_temp4, xmm_temp3);
-  __ movdqu(xmm_temp5, xmm_temp3);
-  __ psrld(xmm_temp2, 1);     // packed left shifting >> 1
-  __ psrld(xmm_temp4, 2);     // packed left shifting >> 2
-  __ psrld(xmm_temp5, 7);     // packed left shifting >> 7
-  __ pxor(xmm_temp2, xmm_temp4);      // xor the shifted versions
-  __ pxor(xmm_temp2, xmm_temp5);
-  __ pxor(xmm_temp2, xmm_temp8);
-  __ pxor(xmm_temp3, xmm_temp2);
-  __ pxor(xmm_temp6, xmm_temp3);      // the result is in xmm6
-
-  __ decrement(blocks);
-  __ jcc(Assembler::zero, L_exit);
-  __ movdqu(xmm_temp0, xmm_temp6);
-  __ addptr(data, 16);
-  __ jmp(L_ghash_loop);
-
-  __ BIND(L_exit);
-  __ pshufb(xmm_temp6, xmm_temp10);          // Byte swap 16-byte result
-  __ movdqu(Address(state, 0), xmm_temp6);   // store the result
   __ leave();
   __ ret(0);
 
@@ -5477,36 +3755,9 @@ void StubGenerator::generate_all() {
   // arraycopy stubs used by compilers
   generate_arraycopy_stubs();
 
-  // don't bother generating these AES intrinsic stubs unless global flag is set
-  if (UseAESIntrinsics) {
-    StubRoutines::x86::_key_shuffle_mask_addr = generate_key_shuffle_mask();  // needed by the others
-    StubRoutines::_aescrypt_encryptBlock = generate_aescrypt_encryptBlock();
-    StubRoutines::_aescrypt_decryptBlock = generate_aescrypt_decryptBlock();
-    StubRoutines::_cipherBlockChaining_encryptAESCrypt = generate_cipherBlockChaining_encryptAESCrypt();
-    if (VM_Version::supports_avx512_vaes() &&  VM_Version::supports_avx512vl() && VM_Version::supports_avx512dq() ) {
-      StubRoutines::_cipherBlockChaining_decryptAESCrypt = generate_cipherBlockChaining_decryptVectorAESCrypt();
-      StubRoutines::_electronicCodeBook_encryptAESCrypt = generate_electronicCodeBook_encryptAESCrypt();
-      StubRoutines::_electronicCodeBook_decryptAESCrypt = generate_electronicCodeBook_decryptAESCrypt();
-      StubRoutines::x86::_counter_mask_addr = counter_mask_addr();
-      StubRoutines::x86::_ghash_poly512_addr = ghash_polynomial512_addr();
-      StubRoutines::x86::_ghash_long_swap_mask_addr = generate_ghash_long_swap_mask();
-      StubRoutines::_galoisCounterMode_AESCrypt = generate_galoisCounterMode_AESCrypt();
-    } else {
-      StubRoutines::_cipherBlockChaining_decryptAESCrypt = generate_cipherBlockChaining_decryptAESCrypt_Parallel();
-    }
-  }
+  generate_aes_stubs();
 
-  if (UseAESCTRIntrinsics) {
-    if (VM_Version::supports_avx512_vaes() && VM_Version::supports_avx512bw() && VM_Version::supports_avx512vl()) {
-      if (StubRoutines::x86::_counter_mask_addr == NULL) {
-        StubRoutines::x86::_counter_mask_addr = counter_mask_addr();
-      }
-      StubRoutines::_counterMode_AESCrypt = generate_counterMode_VectorAESCrypt();
-    } else {
-      StubRoutines::x86::_counter_shuffle_mask_addr = generate_counter_shuffle_mask();
-      StubRoutines::_counterMode_AESCrypt = generate_counterMode_AESCrypt_Parallel();
-    }
-  }
+  generate_ghash_stubs();
 
   if (UseMD5Intrinsics) {
     StubRoutines::_md5_implCompress = generate_md5_implCompress(false, "md5_implCompress");
@@ -5537,22 +3788,6 @@ void StubGenerator::generate_all() {
     StubRoutines::_sha512_implCompress = generate_sha512_implCompress(false, "sha512_implCompress");
     StubRoutines::_sha512_implCompressMB = generate_sha512_implCompress(true, "sha512_implCompressMB");
   }
-
-  // Generate GHASH intrinsics code
-  if (UseGHASHIntrinsics) {
-    if (StubRoutines::x86::_ghash_long_swap_mask_addr == NULL) {
-      StubRoutines::x86::_ghash_long_swap_mask_addr = generate_ghash_long_swap_mask();
-    }
-  StubRoutines::x86::_ghash_byte_swap_mask_addr = generate_ghash_byte_swap_mask();
-    if (VM_Version::supports_avx()) {
-      StubRoutines::x86::_ghash_shuffmask_addr = ghash_shufflemask_addr();
-      StubRoutines::x86::_ghash_poly_addr = ghash_polynomial_addr();
-      StubRoutines::_ghash_processBlocks = generate_avx_ghash_processBlocks();
-    } else {
-      StubRoutines::_ghash_processBlocks = generate_ghash_processBlocks();
-    }
-  }
-
 
   if (UseBASE64Intrinsics) {
     if(VM_Version::supports_avx2() &&
